update:修改decimal类型对应java的double类型

+++ utils/mysqlutil/mysql_utils.go
@@ -19,7 +19,7 @@ func init() {
 	//小数
 	mysqlJavaTypeMap["float"] = "Float"
 	mysqlJavaTypeMap["double"] = "Double"
-	mysqlJavaTypeMap["decimal"] = "BigDecimal"
+	mysqlJavaTypeMap["decimal"] = "Double"
 	//bool
 	mysqlJavaTypeMap["bit"] = "Boolean"
 	//字符串
